fix: pixiConfigのコードの責務を分離する
