+++ virtual_partner_front/src/features/Live2D/utils/pixiConfig.ts
@@ -1,4 +1,4 @@
-import * as PIXI from 'pixi.js';
+import * as PIXI from "pixi.js";
 
 declare global {
   interface Window {
@@ -12,9 +12,6 @@ export const initializePIXI = () => {
   window.PIXI = PIXI;
 };
 
-// リップシンクの制御用の状態
-let isLipSyncEnabled = false;
-
 export const getDefaultPIXIConfig = () => ({
   autoStart: true,
   width: window.innerWidth,
@@ -25,28 +22,3 @@ export const getDefaultPIXIConfig = () => ({
   backgroundAlpha: 0,
   forceCanvas: false,
 });
-
-// アニメーション制御用の関数
-export const controlAnimation = (app: PIXI.Application) => {
-  return {
-    start: () => {
-      if (!app.ticker.started) {
-        app.start();
-      }
-    },
-    stop: () => {
-      app.stop();
-    },
-    startLipSync: () => {
-      isLipSyncEnabled = true;
-      if (!app.ticker.started) {
-        app.start();
-      }
-    },
-    stopLipSync: () => {
-      isLipSyncEnabled = false;
-      app.stop();
-    },
-    isLipSyncActive: () => isLipSyncEnabled,
-  };
-};
